[48] User filter improved
